+++ app/src/main/java/com/example/disastron/MapActivity.java
@@ -37,54 +37,19 @@ import com.google.android.gms.tasks.Task;
 
 public class MapActivity extends MainActivity implements OnMapReadyCallback {
     private GoogleMap mMap;
-    double lat = 12.0,lon = 77.0;
-    int PERMISSION_ID = 44;
-    FusedLocationProviderClient mFusedLocationClient;
+    double lat = 12.0,lon = 77.0;  //default values
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        Log.i("Localine","1");
-        mFusedLocationClient = LocationServices.getFusedLocationProviderClient(this);
-        getLastLocation();
-        Log.i("Localine","2");
         setContentView(R.layout.activity_map);
-        Log.i("Localine","3");
+
         SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager()
                 .findFragmentById(R.id.map);
         mapFragment.getMapAsync(this);
-        Log.i("Localine","4");
-    }
+        Intent intent = getIntent();
+        lat = intent.getDoubleExtra("LAT",0.0);
+        lon = intent.getDoubleExtra("LON",0.0);
 
-
-    @SuppressLint("MissingPermission")
-    private void getLastLocation(){
-        if (checkPermissions()) {
-            if (isLocationEnabled()) {
-                mFusedLocationClient.getLastLocation().addOnCompleteListener(
-                        new OnCompleteListener<Location>() {
-                            @Override
-                            public void onComplete(@NonNull Task<Location> task) {
-                                Location location = task.getResult();
-                                if (location == null) {
-                                    requestNewLocationData();
-                                } else {
-                                    lat = location.getLatitude();
-                                    lon = location.getLongitude();
-                                    Log.i("Location1:",lat+""+lon);
-                                    //latTextView.setText(location.getLatitude()+"");
-                                    //lonTextView.setText(location.getLongitude()+"");
-                                }
-                            }
-                        }
-                );
-            } else {
-                Toast.makeText(this, "Turn on location", Toast.LENGTH_LONG).show();
-                Intent intent = new Intent(Settings.ACTION_LOCATION_SOURCE_SETTINGS);
-                startActivity(intent);
-            }
-        } else {
-            requestPermissions();
-        }
     }
 
     @Override
@@ -92,77 +57,9 @@ public class MapActivity extends MainActivity implements OnMapReadyCallback {
         mMap = googleMap;
         Log.i("Location2:",lat+""+lon);
         LatLng loc = new LatLng(lat, lon);
-        mMap.addMarker(new MarkerOptions().position(loc).title("Your Location"));
-        mMap.moveCamera(CameraUpdateFactory.newLatLng(loc));
-    }
-
-    @SuppressLint("MissingPermission")
-    private void requestNewLocationData(){
-
-        LocationRequest mLocationRequest = new LocationRequest();
-        mLocationRequest.setPriority(LocationRequest.PRIORITY_HIGH_ACCURACY);
-        mLocationRequest.setInterval(0);
-        mLocationRequest.setFastestInterval(0);
-        mLocationRequest.setNumUpdates(1);
-
-        mFusedLocationClient = LocationServices.getFusedLocationProviderClient(this);
-        mFusedLocationClient.requestLocationUpdates(
-                mLocationRequest, mLocationCallback,
-                Looper.myLooper()
-        );
-
-    }
-
-    private LocationCallback mLocationCallback = new LocationCallback() {
-        @Override
-        public void onLocationResult(LocationResult locationResult) {
-            Location mLastLocation = locationResult.getLastLocation();
-            lat = mLastLocation.getLatitude();
-            lon = mLastLocation.getLongitude();
-            //latTextView.setText(mLastLocation.getLatitude()+"");
-            //lonTextView.setText(mLastLocation.getLongitude()+"");
-        }
-    };
-
-    private boolean checkPermissions() {
-        if (ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED &&
-                ActivityCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
-            return true;
-        }
-        return false;
-    }
-
-    private void requestPermissions() {
-        ActivityCompat.requestPermissions(
-                this,
-                new String[]{Manifest.permission.ACCESS_COARSE_LOCATION, Manifest.permission.ACCESS_FINE_LOCATION},
-                PERMISSION_ID
-        );
-    }
-
-    private boolean isLocationEnabled() {
-        LocationManager locationManager = (LocationManager) getSystemService(Context.LOCATION_SERVICE);
-        return locationManager.isProviderEnabled(LocationManager.GPS_PROVIDER) || locationManager.isProviderEnabled(
-                LocationManager.NETWORK_PROVIDER
-        );
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (requestCode == PERMISSION_ID) {
-            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                getLastLocation();
-            }
-        }
-    }
-
-    @Override
-    public void onResume(){
-        super.onResume();
-        if (checkPermissions()) {
-            getLastLocation();
-        }
-
+        //mMap.addMarker(new MarkerOptions().position(loc).title("Your Location"));
+        //mMap.moveCamera(CameraUpdateFactory.newLatLng(loc));
+        mMap.animateCamera(CameraUpdateFactory.newLatLngZoom(loc, 17));
+        mMap.setMyLocationEnabled(true);
     }
 }
